[Android/Fix] 기존 루틴에 없던 기기 등록 시 반영되지 않던 문제 해결

+++ Android/domain/src/main/java/com/hogumiwarts/domain/model/routine/RoutineResult.kt
@@ -20,7 +20,7 @@ sealed class RoutineResult {
 
 
 data class Routine(
-    val routineId: Int,
+    val routineId: Long,
     val routineName: String,
     val routineIcon: String?,
     val gestureName: String?
@@ -53,9 +53,9 @@ data class CommandData(
 )
 
 data class RoutineCreateData(
-    val routineId: Int,
+    val routineId: Long,
     val routineName: String,
     val routineIcon: String,
-    val gestureId: Int,
+    val gestureId: Long,
     val devices: List<CommandDevice>
 )
